+++ src/org/labkey/response/view/MyStudiesResponseServerWebPart.java
@@ -24,11 +24,11 @@ import org.labkey.response.data.MobileAppStudy;
 /**
  * Web part for associating a study short name with a container.
  */
-public class StudyConfigWebPart extends JspView<MobileAppStudy>
+public class MyStudiesResponseServerWebPart extends JspView<MobileAppStudy>
 {
-    public StudyConfigWebPart(ViewContext viewContext)
+    public MyStudiesResponseServerWebPart(ViewContext viewContext)
     {
-        super("/org/labkey/response/view/studySetup.jsp");
+        super("/org/labkey/response/view/MyStudiesResponseServer.jsp");
         setTitle("MyStudies Response Server");
 
         MobileAppStudy bean = ResponseManager.get().getStudy(viewContext.getContainer());
